meson, configure, gtk: fix GDK pkg-config for macOS.
Since commit 7e732caa1e, gdkquartz.h is included directly from
gdkquartz-cocoa-access.h, yet it is in the include/gtk+-3.0/gdk
subdirectory (which is currently not in the dir list to search for
headers). It is made so on purpose so that the headers is also found in
the source tree. Yet it also definitely needs to be found in the install
prefix tree, which is the job of the pkg-config file.