fontchooser: Be more careful about fonts
When we show examples for OpenType features, we use
U+27F6 RIGHTWARDS ARROW, so we should make sure that the
fonts we use have some chance of covering that codepoint.
In the past, we implicitly relied on pangos fontconfig
backend always providing fontsets with 'infinite' coverage.
Other backends may not be so forgiving. Increase our
chances of avoiding hexboxes by appending the generic
sans-serif family at the end.