containers that can be scrolled should have precedence
over child widgets that also react to scrolling
like GtkComboBox, GtkScale and GtkSpinButton,
because otherwise when you're in the middle of
scrolling the view/window you can involuntarily
scroll over the widgets and change its values.

This problem can be seen in applications like
pavucontrol, gnome-tweaks prefs, devhelp prefs,
and so on.

Fixes issue #3092