gsk/gl: add texture library for Glyphy
This adds a new texture library that can upload SDF data from libglyphy
into regions of a texture atlas so that it can be accessed by Glyphy
shaders in the appropriate place and format.

Some of the placement positioning may seem odd in that it needs to follow
a certain format to be decoded from the Glyphy shaders.