gsk: add GskHintNode
The GskHintNode is meant to set flags for the renderer which can be used
to optimize batch submission. Currently, two flags are supported. One will
disable caching of textures from the hint-node to all descendants. The
second will request that the child node be cached as an offscreen texture
and used to blit on subsequent frames.