Add an animated text example
This one is designed to stress font rendering in gsk.
With our current glyph caching, this gets ~20fps on
my system. With glyphy, it gets ~200fps. The glyphy
demo, which just uploads geometry once and keeps it
on the gpu, gets ~1800fps.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9714>